// src/types/mock-type/mock-data.ts
type MockData = {
  name: string[];
  posterImage: string[];
  backgroundImage: string[];
  backgroundColor: string[];
  description: string[];
  rating: number[];
  scoresCount: number[];
  director: string[];
  starring: string[];
  runTime: number[];
  genre: string[];
  released: number[];
  userName: string[];
  email: string[];
  password: string[];
  avatar?: string[];
  publishDate: string[];
  previewVideoLink: string[];
  videoLink: string[];
}

export type { MockData };